color: Split most of the color parsing to a separate crate.
