+++ Cargo.toml
@@ -34,4 +34,4 @@ dummy_match_byte = []
 skip_long_tests = []
 
 [workspace]
-members = [".", "./macros"]
+members = [".", "./macros", "./color"]
